+++ CentralStation/Framework/Constants/MapSubActions.cs
@@ -6,6 +6,9 @@ internal class MapSubActions
     /// <summary>The sub-action which opens the bookshelf dialogue.</summary>
     public const string Bookshelf = "Bookshelf";
 
+    /// <summary>The sub-action which shows the dialogue to buy a Joja Cola.</summary>
+    public const string ColaMachine = "ColaMachine";
+
     /// <summary>The sub-action which opens a destination menu through the main Central Station ticket booth, with a rare chance of showing a message first.</summary>
     public const string TicketBooth = "TicketBooth";