+++ CentralStation/docs/release-notes.md
@@ -6,6 +6,7 @@
   - boat, bus, and train networks.
   - Central Station map and custom ticket machine sprite [commissioned from Kisaa](https://next.nexusmods.com/profile/crystalinerose) (thanks!).
   - food court, gift shop, tourists, and interactive bookshelves.
+  - rare interactions in the Central Station.
   - integrations with the Bus Locations, CJB Cheats Menu, and Train Station mods.
   - data assets to register stops, tourists, and bookshelf messages through Content Patcher.
   - C# mod API to register stops.